+++ src/components/RedPacket/RedPacketEnvelope.tsx
@@ -2,7 +2,7 @@
 import { type HTMLProps } from 'react';
 
 import { Image } from '@/components/Image.js';
-import { type RedPacketCoverOptions, useRedPacketCover } from '@/mask/plugins/red-packet/hooks/useRedPacketCover.js';
+import { type RedPacketCoverOptions, useRedPacketCover } from '@/components/RedPacket/hooks/useRedPacketCover.js';
 
 interface Props extends HTMLProps<HTMLDivElement>, RedPacketCoverOptions {}
