+++ src/components/RedPacket/hooks/useAvailability.ts
