+++ src/components/RedPacket/hooks/useCreateCallback.tsx
@@ -17,7 +17,7 @@ import { type Hex, keccak256 } from 'viem';
 import { useChainContext } from '@/hooks/useChainContext.js';
 import type { HappyRedPacketV4 } from '@/mask/bindings/constants.js';
 import { EVMWeb3 } from '@/mask/bindings/index.js';
-import { useRedPacketContract } from '@/mask/plugins/red-packet/hooks/useRedPacketContract.js';
+import { useRedPacketContract } from '@/components/RedPacket/hooks/useRedPacketContract.js';
 
 export interface RedPacketSettings {
     shares: number;
