+++ src/components/RedPacket/hooks/usePlatformType.ts
