+++ src/locales/en/messages.po
@@ -337,6 +337,9 @@ msgstr "Trending"
 msgid "Log out @{0}"
 msgstr "Log out @{0}"
 
+msgid "DAOs"
+msgstr "DAOs"
+
 msgid "MAX"
 msgstr "MAX"
 
@@ -1404,9 +1407,6 @@ msgstr "This token contract is open source. You can check the contract code for 
 
 msgid "Darkweb transactions found."
 msgstr "Darkweb transactions found."
-
-msgid "DAO"
-msgstr "DAO"
 
 msgid "{0} Comments"
 msgstr "{0} Comments"
