+++ src/modals/RedPacketModal/HistoryView.tsx
@@ -7,7 +7,7 @@ import { VirtualList } from '@/components/VirtualList/VirtualList.js';
 import { VirtualListFooter } from '@/components/VirtualList/VirtualListFooter.js';
 import { ScrollListKey } from '@/constants/enum.js';
 import { useChainContext } from '@/hooks/useChainContext.js';
-import { useRedPacketHistory } from '@/mask/plugins/red-packet/hooks/useRedPacketHistory.js';
+import { useRedPacketHistory } from '@/components/RedPacket/hooks/useRedPacketHistory.js';
 import { RedPacketDetailItem } from '@/modals/RedPacketModal/RedPacketDetailItem.js';
 import { FireflyRedPacketAPI } from '@/providers/red-packet/types.js';
